requirements.txt
# Core Testing 
pytest==8.4.1
pytest-html==4.1.1
pytest-xdist==3.8.0
pytest-cov==5.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# AWS/NCP SDK 
boto3==1.34.144
botocore==1.34.144

# HTTP & API
requests==2.32.5

# Configuration
python-dotenv==1.1.1
PyYAML==6.0.1

# Logging
coloredlogs==15.0.1

# Phase 2: Image Processing
Pillow==10.4.0
pillow-heif==0.18.0

# Phase 2: Monitoring & Dashboard
pandas==2.2.2
streamlit==1.38.0
plotly==5.24.0

# Phase 2: Utilities
tenacity==8.5.0
tqdm==4.66.5
faker==28.4.1

# Optional: Async HTTP 
# aiohttp==3.10.5